Sample phenological metrics dataset

Data frame with the phenological metrics extracted from the object cf, created following the example provided in the extract_pheno() documentation. It is used in the documentation of functions assign_season() and aggregate_pheno().

data(dt_pheno)

Format

A data table with 3 records (identified cycles) and the following fields:

id, year, cycle, begin, end, maxval, weight: inherited from dt_cycles

sos, eos, los, pop, mgs, rsp, rau, peak, msp, mau: phenological metrics created by method "trs" (see phenopix::PhenoTrs()).

Examples

data(dt_pheno)
dt_pheno
#>    id year cycle      begin        end     maxval   weight        sos
#> 1:  1 2020     1 2020-04-27 2020-10-14 2020-07-13 88.71541 2020-05-30
#> 2:  2 2020     1 2020-01-04 2020-07-15 2020-04-21 92.78070 2020-03-14
#> 3:  2 2020     2 2020-07-16 2020-10-15 2020-08-08 13.91427 2020-07-29
#>           eos los        pop       mgs rsp rau      peak       msp       mau
#> 1: 2020-08-26  88 2020-07-01 0.8183139  NA  NA 0.8885069 0.4805396 0.4575946
#> 2: 2020-06-02  80 2020-05-01 0.7291655  NA  NA 0.8196339 0.5021371 0.5032177
#> 3: 2020-08-20  22 2020-08-09 0.4262500  NA  NA 0.4886998 0.3193936 0.2943121