Paths of the rasters used in documentation examples

The sample dataset contains Sentinel-2 derived images (NDVI index and SCL) used in the documentation examples of extract_s2ts(). This archive is not provided as a raster stack using standard data() way because extract_s2ts() accepts paths as input.

sample_paths(prod = "NDVI")

Arguments

prod
"NDVI" or "SCL".

Format

Each sample archive include 60 images acquired between 2020-01-01 and 2020-09-30 covering the sampleroi extent, provided in the sen2r output format. NDVI images are saved in integer format (values are scaled using a 10^4 factor).

Data size is very small (13x4 pixels per each image) in order not to be heavy in terms of data size and to ensure examples to be rapidly runnable. The vignette "The sen2rts workflow" makes use of a larger dataset in order to show a more plausible use case.

Value

sample_paths() returns the paths of the sample archive (NDVI or SCL, depending on argument prod).

Examples

sen2r_ndvi_paths <- sample_paths("NDVI")
head(sen2r_ndvi_paths)

#> [1] "/tmp/RtmpDtevJN/temp_libpath429b31237ea4/sen2rts/extdata/sen2r/sampleroi/NDVI/S2B2A_20200104_022_sampleroi_NDVI_10.tif"
#> [2] "/tmp/RtmpDtevJN/temp_libpath429b31237ea4/sen2rts/extdata/sen2r/sampleroi/NDVI/S2A2A_20200109_022_sampleroi_NDVI_10.tif"
#> [3] "/tmp/RtmpDtevJN/temp_libpath429b31237ea4/sen2rts/extdata/sen2r/sampleroi/NDVI/S2B2A_20200114_022_sampleroi_NDVI_10.tif"
#> [4] "/tmp/RtmpDtevJN/temp_libpath429b31237ea4/sen2rts/extdata/sen2r/sampleroi/NDVI/S2A2A_20200119_022_sampleroi_NDVI_10.tif"
#> [5] "/tmp/RtmpDtevJN/temp_libpath429b31237ea4/sen2rts/extdata/sen2r/sampleroi/NDVI/S2B2A_20200124_022_sampleroi_NDVI_10.tif"
#> [6] "/tmp/RtmpDtevJN/temp_libpath429b31237ea4/sen2rts/extdata/sen2r/sampleroi/NDVI/S2A2A_20200129_022_sampleroi_NDVI_10.tif"

sen2r_scl_paths <- sample_paths("SCL")
head(sen2r_scl_paths)

#> [1] "/tmp/RtmpDtevJN/temp_libpath429b31237ea4/sen2rts/extdata/sen2r/sampleroi/SCL/S2B2A_20200104_022_sampleroi_SCL_10.tif"
#> [2] "/tmp/RtmpDtevJN/temp_libpath429b31237ea4/sen2rts/extdata/sen2r/sampleroi/SCL/S2A2A_20200109_022_sampleroi_SCL_10.tif"
#> [3] "/tmp/RtmpDtevJN/temp_libpath429b31237ea4/sen2rts/extdata/sen2r/sampleroi/SCL/S2B2A_20200114_022_sampleroi_SCL_10.tif"
#> [4] "/tmp/RtmpDtevJN/temp_libpath429b31237ea4/sen2rts/extdata/sen2r/sampleroi/SCL/S2A2A_20200119_022_sampleroi_SCL_10.tif"
#> [5] "/tmp/RtmpDtevJN/temp_libpath429b31237ea4/sen2rts/extdata/sen2r/sampleroi/SCL/S2B2A_20200124_022_sampleroi_SCL_10.tif"
#> [6] "/tmp/RtmpDtevJN/temp_libpath429b31237ea4/sen2rts/extdata/sen2r/sampleroi/SCL/S2A2A_20200129_022_sampleroi_SCL_10.tif"